Create Zookeeper
This project  helps you understand some syntax basics and learn how to work with variables, data storage types such as lists, and while loops. 
The end goal of this project was to write a function that lets you visit every animal with the number of the habitat and continuously show the animals without having to be restarted.